Add SFTPStreamTest::testFopenFcloseCreatesFile()
